Restructure code: Split code into logical classes
